[refactor] moe: compute softmax after argmax
According to @TimDettmers:

So the flow from the balancing loss to the gate is the same. The
difference is when we multiply the softmax probability by the
expert we have a softmax gradient flow from the expert to the gate.
The problem with that is that (1) softmax gradients are small since
the gradient is p*(1-p), (2) if the gradient is close to 0 then the
deltas/errors coming from the expert are multiplied by this value on
the way to the gate. This means that a probability close to zero
should be avoided or otherwise errors from the 2nd expert will not
update the gate and the gate "does not know about the improvement to
the expert". Adam will help with this expect a little bit, since the
running differences for such an expert are small but we only get on
average every k/num_experts samples an running differences update
for adam. So adam might not be able to adjust quickly enough for a
situation like this.

The second piece is since we normalize in the softmax by the sum of
all exponentiated terms, and each term is positive, the normalizing
terms grows with the size of the number of experts. So if we have
1024 experts, it would be very likely that most of the probability
terms are close to zero and hence have very small gradients. If we
just take the top-k values and normalize them, then the normalizing
term is much smaller and the minimum probability is still sizable to
allow gradient flow from expert to gate.